Add files via upload
I tested them and they work with the isotropic elasticity tensor, but the solver won't converge with the Ti64's elasticity tensor. Also, I have added the
comments as per my understanding. Hope they will be helpful for you to debug the issues or to use them directly in making changes in your code.